Create quantum/QFT context, transfer/update old QFT theorems
This commit includes the formal creation of the quantum/QFT context (the QFT
folder was there from before, but the new 'context' had not yet been created).
Updated the InverseFourierTransform class in
quantum/QFT/fourierTransformOps.py
and transferred and updated the old QFT/theorems.ipynb over to the new
context's QFT/_theorems_.ipynb. Transfer/update still uses older Sqrt class
instead of the recently-created sqrt() fxn because the sqrt() fxn work is still
in another branch that hasn't been merged yet. Related Issue: #97.
