+++ packages/proveit/physics/quantum/QFT/_brief_description_.txt
@@ -1,0 +1,1 @@
+For work related to Nielsen & Chuang's quantum fourier transform (QFT) alg'm
