+++ packages/proveit/physics/quantum/_sub_contexts_.txt
@@ -1,1 +1,2 @@
+QFT
 QPE
